COMMERCE-4881 move
